HTML-Handbook-v2
New version of the HTML-handbook.

The jar file create a json file at the top level containing all the structure of the handbook. This help to display faster large handbook. The structure does not need to be recalculate during each refresh. Moreover, the structure is calculated by the jar file and does not require to be manually maintained.
Source of the jar are contained in the "ContentFolder-src" folder. It is not required to download it to use the tool.

Some html tags did change.